requirements.txt
arrow==1.2.3
asgiref==3.7.2
binaryornot==0.4.4
certifi==2023.7.22
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cookiecutter==2.2.3
crispy-tailwind==0.5.0
Django==4.2.3
django-crispy-forms==2.0
django-taggit==4.0.0
django-tailwind==3.6.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
psycopg2==2.9.6
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
urllib3==2.0.4
